Rename the LLMs store
