+++ src/common/state/store-chats.ts
@@ -3,7 +3,7 @@ import { devtools, persist } from 'zustand/middleware';
 import { v4 as uuidv4 } from 'uuid';
 
 import { DLLMId } from '~/modules/llms/llm.types';
-import { useModelsStore } from '~/modules/llms/llm.store';
+import { useModelsStore } from '~/modules/llms/store-llms';
 
 import { countModelTokens } from '../llm-util/token-counter';
 import { defaultSystemPurposeId, SystemPurposeId } from '../../data';
