+++ src/modules/llms/configurator/LLMList.tsx
@@ -5,7 +5,7 @@ import { List, Sheet, Typography } from '@mui/joy';
 
 import { LLMListItem } from './LLMListItem';
 import { findVendorById } from '../vendor.registry';
-import { useModelsStore } from '../llm.store';
+import { useModelsStore } from '../store-llms';
 
 
 export function LLMList() {
